Implemented Game Loop, Ball-Ball Collision
Added a Game Loop, however the usual Update Model step occurs concurrently with Rendering when using Game Arena, so just an Update method is called and any changes to the billiard positions are automatically rendered as they are calculated. Pockets match to the table ratio, I will soon add the table+billiards to scale with the resizable window. Ball on ball collision detection and change of velocity has been implemented also. Ball on wall collision will be added next, followed by pocketing of the billiards.